+++ ExilenceClient/src/app/authorize/dashboard/dashboard.component.ts
@@ -10,6 +10,7 @@ import { InfoDialogComponent } from '../components/info-dialog/info-dialog.compo
 import { MatDialog } from '@angular/material';
 import { SettingsService } from '../../shared/providers/settings.service';
 import { Subscription } from 'rxjs/internal/Subscription';
+import { IncomeService } from '../../shared/providers/income.service';
 
 @Component({
   selector: 'app-dashboard',
@@ -31,7 +32,8 @@ export class DashboardComponent implements OnInit, OnDestroy, AfterViewInit {
     private analyticsService: AnalyticsService,
     private settingsService: SettingsService,
     private router: Router,
-    private dialog: MatDialog
+    private dialog: MatDialog,
+    private incomeService: IncomeService
   ) {
 
     this.partyService.recentParties.subscribe(parties => {
@@ -97,6 +99,7 @@ export class DashboardComponent implements OnInit, OnDestroy, AfterViewInit {
       this.partyService.leaveParty(this.partyService.party.name, this.player);
       setTimeout(() => {
         this.partyService.joinParty(partyName, this.player);
+        this.incomeService.Snapshot();
         this.partyService.addPartyToRecent(partyName);
         this.router.navigateByUrl('/404', { skipLocationChange: true }).then(() =>
           this.router.navigate(['/authorized/party']));