+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.8.13] - 2018-12-XX
+## [0.9.0] - 2018-12-19
 - Added support for pricing of scarabs
+- Added support to price linked items
+- Added support to price 6-socketed items
+- Added support to price elder/shaper items
+- Added support for pricing gems (with level and quality)
+- Added support to price basetypes (with and without elder/shaper)
+- Added new column 'links' to net worth table
+- Added new column 'quality' to net worth table
+- Pricing of items should now be more accurate
 - Added a new, simplified launch-screen for existing users
 - Added persistance between restarts to mask-setting for groupnames
+- Fixed a bug where gain for the group summary was not updated properly
 
 ## [0.8.12] - 2018-12-18
 - Optimized memory usage by dropping disposing of unused objects